requirements.txt
django==4.1.13
djangorestframework==3.14.0
gunicorn==20.1.0
uvicorn==0.20.0
celery==5.2.7
django_environ==0.9.0
mysqlclient==2.2.1
drf_yasg==1.21.4
redis==4.5.1
graypy==2.1.0

pyjwt==2.6.0
cryptography==39.0.1

mypy==1.0.1
model-bakery==1.10.1
requests-mock==1.10.0